add check for puppet3 or puppet4 to determine which groups the nagios user is added to.
In puppet3, the nagios user is set as a member of the puppet group.

In puppet4, the puppet group is not created, except on nodes that we are cutting over from puppet3 to 4.
Therefore we don't try to put nagios user in this group so puppet doesn't throw an error.

Change-Id: I2678809864128c83ec0e6f8f09fcc55853246520