fix: add environment to puppetdb query
Merely using tag is not enough - E.g. resources that are declared in a
'melbourne' module will be automatically tagged with 'melbourne' even
though they are in 'nectar' environment. This leads to melbourne
naginator script pulling in any resources declared in 'melbourne'
module, regardless of the environment the resource is created.

In previous versions of puppet, filtering by 'environment' didn't work.
It now does. So now we can replace 'tag' with 'environment'.

In our puppet code, nagios resources have been tagged with
$config_environment to work around the filtering bug. This tag can now
be removed after this change goes in.

Change-Id: Ic85d14b8741c2401c98b9275c1202eecc9f86a47